pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "django-fileindex"
version = "0.2.0"
description = "A Django app for file deduplication and indexing using SHA hashes"
readme = "README.md"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.1",
]
requires-python = ">=3.10"
dependencies = [
    "Django>=4.2",
    "Pillow>=10.0.0",
    "pillow-avif-plugin>=1.3.0",
    "thumbhash-python>=1.0.1",
    "django-pg-queue>=0.8.0",
    "django-sendfile2>=0.7.0",
    "watchdog>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/django-fileindex"
Repository = "https://github.com/yourusername/django-fileindex"
Issues = "https://github.com/yourusername/django-fileindex/issues"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-django>=4.5.0",
    "pytest-cov>=5.0.0",
    "coverage[toml]>=7.0.0",
    "black>=24.0.0",
    "ruff>=0.5.0",
    "psycopg2-binary>=2.9.0",
]

[tool.hatch.build.targets.wheel]
packages = ["fileindex"]

[tool.coverage.run]
source = ["fileindex"]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/test_*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
python_files = ["test_*.py", "*_test.py", "tests.py"]
testpaths = ["tests", "fileindex"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "postgres: marks tests that require PostgreSQL",
    "sqlite: marks tests that can run with SQLite",
]

[tool.ruff]
line-length = 88
target-version = "py310"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]